@@ -50,93 +61,93 @@
   },
   "homepage": "https://github.com/mattlewis92/angular-calendar#readme",
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/build-ng-packagr": "^0.13.9",
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.9",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@commitlint/cli": "^7.6.1",
-    "@commitlint/config-conventional": "^7.6.0",
-    "@commitlint/prompt": "^7.6.1",
-    "@compodoc/compodoc": "^1.1.9",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.3",
+    "@angular-devkit/build-angular": "^0.802.2",
+    "@angular-devkit/build-ng-packagr": "^0.802.2",
+    "@angular/animations": "^8.2.2",
+    "@angular/cdk": "^8.1.3",
+    "@angular/cli": "^8.2.2",
+    "@angular/common": "^8.2.2",
+    "@angular/compiler": "^8.2.2",
+    "@angular/compiler-cli": "^8.2.2",
+    "@angular/core": "^8.2.2",
+    "@angular/forms": "^8.2.2",
+    "@angular/language-service": "^8.2.2",
+    "@angular/platform-browser": "^8.2.2",
+    "@angular/platform-browser-dynamic": "^8.2.2",
+    "@angular/router": "^8.2.2",
+    "@commitlint/cli": "^8.1.0",
+    "@commitlint/config-conventional": "^8.1.0",
+    "@commitlint/prompt": "^8.1.0",
+    "@compodoc/compodoc": "^1.1.10",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.0",
     "@stackblitz/sdk": "^1.3.0",
-    "@types/chai": "^4.1.7",
-    "@types/mocha": "^5.2.6",
+    "@types/chai": "^4.2.0",
+    "@types/mocha": "^5.2.7",
     "@types/moment-timezone": "^0.5.12",
-    "@types/node": "^11.13.11",
-    "@types/sinon": "^7.0.11",
-    "@types/sinon-chai": "^3.2.2",
+    "@types/node": "^12.7.2",
+    "@types/sinon": "^7.0.13",
+    "@types/sinon-chai": "^3.2.3",
     "angular-cli-ghpages": "^0.5.3",
-    "angulartics2": "^7.5.2",
+    "angulartics2": "^8.0.0",
     "angularx-flatpickr": "^6.1.0",
-    "autoprefixer": "^9.5.1",
+    "autoprefixer": "^9.6.1",
     "bootstrap-css-only": "^4.3.1",
     "chai": "^4.2.0",
     "codecov": "^3.5.0",
-    "codelyzer": "^5.0.1",
-    "commitizen": "^3.1.1",
-    "copyfiles": "^2.1.0",
-    "core-js": "^2.6.6",
-    "css-loader": "^2.1.1",
+    "codelyzer": "^5.1.0",
+    "commitizen": "^4.0.3",
+    "copyfiles": "^2.1.1",
+    "core-js": "^3.2.1",
+    "cp-cli": "^2.0.0",
+    "cross-env": "^5.2.0",
+    "css-loader": "^3.2.0",
     "date-fns": "1.29.0",
-    "flatpickr": "^4.5.7",
+    "del-cli": "^2.0.0",
+    "flatpickr": "^4.6.2",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "highlightjs-loader": "^0.2.3",
-    "husky": "^2.3.0",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "husky": "^3.0.4",
+    "karma": "^4.2.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.5",
-    "mocha": "^6.1.4",
+    "mocha": "^6.2.0",
     "moment": "^2.24.0",
-    "moment-timezone": "^0.5.25",
-    "ng-packagr": "^4.7.1",
+    "moment-timezone": "^0.5.26",
+    "ng-packagr": "^5.5.0",
+    "ngx-clipboard": "^12.2.0",
     "ngx-contextmenu": "^5.2.0",
-    "ngx-perfect-scrollbar": "^7.2.1",
+    "ngx-perfect-scrollbar": "^8.0.0",
     "node-sass": "^4.12.0",
-    "postcss-cli": "^6.1.2",
+    "npm-run-all": "^4.1.5",
+    "postcss-cli": "^6.1.3",
     "postcss-flexibility": "^2.0.0",
-    "prettier": "^1.17.1",
-    "pretty-quick": "^1.11.0",
-    "rrule": "^2.6.0",
+    "prettier": "^1.18.2",
+    "pretty-quick": "^1.11.1",
+    "rrule": "^2.6.2",
     "rxjs": "^6.5.2",
-    "sinon": "^7.3.2",
+    "sinon": "^7.4.1",
     "sinon-chai": "^3.3.0",
-    "standard-version": "^5.0.2",
-    "stylelint": "^10.0.1",
+    "standard-version": "^7.0.0",
+    "stylelint": "^10.1.0",
     "stylelint-config-prettier": "^5.2.0",
     "stylelint-config-standard": "^18.3.0",
-    "ts-node": "^8.1.0",
-    "tsickle": "0.34.0",
-    "tslib": "^1.9.0",
-    "tslint": "^5.16.0",
+    "ts-node": "^8.3.0",
+    "tsickle": "^0.36.0",
+    "tslib": "^1.10.0",
+    "tslint": "^5.18.0",
     "tslint-config-mwl": "^0.6.0",
-    "typescript": "~3.2.4",
-    "zone.js": "^0.8.29"
+    "typescript": "~3.5.3",
+    "zone.js": "~0.9.1"
   },
-  "peerDependencies": {
-    "@angular/animations": ">=6.0.0 <9.0.0",
-    "@angular/common": ">=6.0.0 <9.0.0",
-    "@angular/core": ">=6.0.0 <9.0.0"
-  },
+  "peerDependencies": {},
   "dependencies": {
-    "angular-draggable-droppable": "^4.3.0",
+    "angular-draggable-droppable": "^4.3.2",
     "angular-resizable-element": "^3.2.4",
-    "calendar-utils": "^0.3.1",
-    "positioning": "^1.4.0"
+    "calendar-utils": "^0.3.3",
+    "positioning": "^2.0.0"
   },
   "sideEffects": [
     "*.css",
