@@ -45,7 +45,7 @@ const colors: any = {
   templateUrl: 'template.html'
 })
 export class DemoComponent {
-  @ViewChild('modalContent') modalContent: TemplateRef<any>;
+  @ViewChild('modalContent', { static: true }) modalContent: TemplateRef<any>;
 
   view: CalendarView = CalendarView.Month;
 
@@ -123,7 +123,6 @@ export class DemoComponent {
 
   dayClicked({ date, events }: { date: Date; events: CalendarEvent[] }): void {
     if (isSameMonth(date, this.viewDate)) {
-      this.viewDate = date;
       if (
         (isSameDay(this.viewDate, date) && this.activeDayIsOpen === true) ||
         events.length === 0
@@ -132,6 +131,7 @@ export class DemoComponent {
       } else {
         this.activeDayIsOpen = true;
       }
+      this.viewDate = date;
     }
   }
 
