@@ -4,7 +4,7 @@ const postCssFlexibility = require('postcss-flexibility');
 module.exports = () => ({
   plugins: [
     autoprefixer({
-      browsers: [
+      overrideBrowserslist: [
         '> 1%',
         'last 4 versions',
         'last 20 Chrome versions',
