@@ -1,7 +1,5 @@
-import { isInside } from './util';
+import { isInside, isWithinThreshold } from './util';
 import { ValidateDragParams } from 'angular-draggable-droppable';
-
-const DRAG_THRESHOLD = 1;
 
 export class CalendarDragHelper {
   private readonly startPosition: ClientRect;
@@ -26,9 +24,6 @@ export class CalendarDragHelper {
     dragAlreadyMoved: boolean;
     transform: ValidateDragParams['transform'];
   }): boolean {
-    const isWithinThreshold =
-      Math.abs(x) > DRAG_THRESHOLD || Math.abs(y) > DRAG_THRESHOLD;
-
     if (snapDraggedEvents) {
       const newRect: ClientRect = Object.assign({}, this.startPosition, {
         left: this.startPosition.left + transform.x,
@@ -38,11 +33,11 @@ export class CalendarDragHelper {
       });
 
       return (
-        (isWithinThreshold || dragAlreadyMoved) &&
+        (isWithinThreshold({ x, y }) || dragAlreadyMoved) &&
         isInside(this.dragContainerElement.getBoundingClientRect(), newRect)
       );
     } else {
-      return isWithinThreshold || dragAlreadyMoved;
+      return isWithinThreshold({ x, y }) || dragAlreadyMoved;
     }
   }
 }
