@@ -12,7 +12,9 @@ import {
   ComponentFactory,
   Inject,
   Renderer2,
-  TemplateRef
+  TemplateRef,
+  OnChanges,
+  SimpleChanges
 } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
 import { PlacementArray, positionElements } from 'positioning';
@@ -58,7 +60,7 @@ export class CalendarTooltipWindowComponent {
 @Directive({
   selector: '[mwlCalendarTooltip]'
 })
-export class CalendarTooltipDirective implements OnDestroy {
+export class CalendarTooltipDirective implements OnDestroy, OnChanges {
   @Input('mwlCalendarTooltip') contents: string; // tslint:disable-line no-input-rename
 
   @Input('tooltipPlacement') placement: PlacementArray = 'auto'; // tslint:disable-line no-input-rename
@@ -86,6 +88,18 @@ export class CalendarTooltipDirective implements OnDestroy {
     this.tooltipFactory = componentFactoryResolver.resolveComponentFactory(
       CalendarTooltipWindowComponent
     );
+  }
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (
+      this.tooltipRef &&
+      (changes.contents || changes.customTemplate || changes.event)
+    ) {
+      this.tooltipRef.instance.contents = this.contents;
+      this.tooltipRef.instance.customTemplate = this.customTemplate;
+      this.tooltipRef.instance.event = this.event;
+      this.tooltipRef.changeDetectorRef.markForCheck();
+    }
   }
 
   ngOnDestroy(): void {
@@ -136,7 +150,7 @@ export class CalendarTooltipDirective implements OnDestroy {
     this.cancelTooltipDelay$.next();
   }
 
-  private positionTooltip(previousPosition?: string): void {
+  private positionTooltip(previousPositions: string[] = []): void {
     if (this.tooltipRef) {
       this.tooltipRef.changeDetectorRef.detectChanges();
       this.tooltipRef.instance.placement = positionElements(
@@ -146,8 +160,13 @@ export class CalendarTooltipDirective implements OnDestroy {
         this.appendToBody
       );
       // keep re-positioning the tooltip until the arrow position doesn't make a difference
-      if (previousPosition !== this.tooltipRef.instance.placement) {
-        this.positionTooltip(this.tooltipRef.instance.placement);
+      if (
+        previousPositions.indexOf(this.tooltipRef.instance.placement) === -1
+      ) {
+        this.positionTooltip([
+          ...previousPositions,
+          this.tooltipRef.instance.placement
+        ]);
       }
     }
   }
