@@ -8,22 +8,35 @@ import {
 import {
   trigger,
   style,
+  state,
   transition,
   animate,
   AnimationTriggerMetadata
 } from '@angular/animations';
 import { CalendarEvent } from 'calendar-utils';
-import { trackByEventId } from '../common/util';
+import { isWithinThreshold, trackByEventId } from '../common/util';
 
 export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
-  transition('void => *', [
-    style({ height: 0, overflow: 'hidden' }),
-    animate('150ms', style({ height: '*' }))
-  ]),
-  transition('* => void', [
-    style({ height: '*', overflow: 'hidden' }),
-    animate('150ms', style({ height: 0 }))
-  ])
+  state(
+    'void',
+    style({
+      height: 0,
+      overflow: 'hidden',
+      'padding-top': 0,
+      'padding-bottom': 0
+    })
+  ),
+  state(
+    '*',
+    style({
+      height: '*',
+      overflow: 'hidden',
+      'padding-top': '*',
+      'padding-bottom': '*'
+    })
+  ),
+  transition('* => void', animate('150ms ease-out')),
+  transition('void => *', animate('150ms ease-in'))
 ]);
 
 @Component({
@@ -35,6 +48,7 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
       let-eventClicked="eventClicked"
       let-isOpen="isOpen"
       let-trackByEventId="trackByEventId"
+      let-validateDrag="validateDrag"
     >
       <div class="cal-open-day-events" [@collapse] *ngIf="isOpen">
         <div
@@ -45,6 +59,7 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
           dragActiveClass="cal-drag-active"
           [dropData]="{ event: event }"
           [dragAxis]="{ x: event.draggable, y: event.draggable }"
+          [validateDrag]="validateDrag"
         >
           <span
             class="cal-event"
@@ -74,7 +89,8 @@ export const collapseAnimation: AnimationTriggerMetadata = trigger('collapse', [
         events: events,
         eventClicked: eventClicked,
         isOpen: isOpen,
-        trackByEventId: trackByEventId
+        trackByEventId: trackByEventId,
+        validateDrag: validateDrag
       }"
     >
     </ng-template>
@@ -98,4 +114,6 @@ export class CalendarOpenDayEventsComponent {
   }>();
 
   trackByEventId = trackByEventId;
+
+  validateDrag = isWithinThreshold;
 }
