@@ -4,6 +4,7 @@ import { map, take, filter } from 'rxjs/operators';
 import StackBlitzSDK from '@stackblitz/sdk';
 import { Angulartics2GoogleAnalytics } from 'angulartics2/ga';
 import { sources as demoUtilsSources } from './demo-modules/demo-utils/sources';
+import { Subject } from 'rxjs';
 
 interface Source {
   filename: string;
@@ -89,6 +90,7 @@ export class DemoAppComponent implements OnInit {
   isMenuVisible = false;
   firstDemoLoaded = false;
   searchText = '';
+  copied$ = new Subject<boolean>();
 
   constructor(private router: Router, analytics: Angulartics2GoogleAnalytics) {
     analytics.startTracking();
@@ -147,18 +149,10 @@ export class DemoAppComponent implements OnInit {
       [path: string]: string;
     } = {
       'index.html': `
-<link href="https://unpkg.com/bootstrap-css-only@${
-        dependencyVersions.bootstrap
-      }/css/bootstrap.min.css" rel="stylesheet">
-<link href="https://unpkg.com/font-awesome@${
-        dependencyVersions.fontAwesome
-      }/css/font-awesome.css" rel="stylesheet">
-<link href="https://unpkg.com/angular-calendar@${
-        dependencyVersions.angularCalendar
-      }/css/angular-calendar.css" rel="stylesheet">
-<link href="https://unpkg.com/flatpickr@${
-        dependencyVersions.flatpickr
-      }/dist/flatpickr.css" rel="stylesheet">
+<link href="https://unpkg.com/bootstrap-css-only@${dependencyVersions.bootstrap}/css/bootstrap.min.css" rel="stylesheet">
+<link href="https://unpkg.com/font-awesome@${dependencyVersions.fontAwesome}/css/font-awesome.css" rel="stylesheet">
+<link href="https://unpkg.com/angular-calendar@${dependencyVersions.angularCalendar}/css/angular-calendar.css" rel="stylesheet">
+<link href="https://unpkg.com/flatpickr@${dependencyVersions.flatpickr}/dist/flatpickr.css" rel="stylesheet">
 <mwl-demo-component>Loading...</mwl-demo-component>
 `.trim(),
       'main.ts': `
@@ -220,15 +214,11 @@ platformBrowserDynamic().bootstrapModule(BootstrapModule).then(ref => {
           '@angular/animations': dependencyVersions.angular,
           rxjs: dependencyVersions.rxjs,
           'zone.js': dependencyVersions.zoneJs,
-          'angular-draggable-droppable': `^${
-            dependencyVersions.angularDraggableDroppable
-          }`,
-          'angular-resizable-element': `^${
-            dependencyVersions.angularResizableElement
-          }`,
+          'angular-draggable-droppable': `^${dependencyVersions.angularDraggableDroppable}`,
+          'angular-resizable-element': `^${dependencyVersions.angularResizableElement}`,
           'date-fns': dependencyVersions.dateFns,
           'angular-calendar': dependencyVersions.angularCalendar,
-          '@ng-bootstrap/ng-bootstrap': dependencyVersions.ngBootstrap,
+          '@ng-bootstrap/ng-bootstrap': '5.0.0', // pinned due to issue with stackblitz generation
           rrule: dependencyVersions.rrule,
           'calendar-utils': dependencyVersions.calendarUtils,
           flatpickr: dependencyVersions.flatpickr,
@@ -239,5 +229,12 @@ platformBrowserDynamic().bootstrapModule(BootstrapModule).then(ref => {
         openFile: 'demo/component.ts'
       }
     );
+  }
+
+  copied() {
+    this.copied$.next(true);
+    setTimeout(() => {
+      this.copied$.next(false);
+    }, 1000);
   }
 }
